Add the no-eth1-sim, refactor sim
