+++ .github/workflows/test-suite.yml
@@ -49,12 +49,21 @@ jobs:
     - uses: actions/checkout@v1
     - name: Build the root Dockerfile
       run: docker build .
-  simulator-ubuntu:
+  eth1-simulator-ubuntu:
     runs-on: ubuntu-latest
     needs: cargo-fmt
     steps:
     - uses: actions/checkout@v1
     - name: Install ganache-cli
       run: sudo npm install -g ganache-cli
-    - name: Run the beacon chain sim
-      run: cargo run --release --bin simulator beacon-chain-sim
+    - name: Run the beacon chain sim that starts from an eth1 contract
+      run: cargo run --release --bin simulator eth1-sim
+  no-eth1-simulator-ubuntu:
+    runs-on: ubuntu-latest
+    needs: cargo-fmt
+    steps:
+    - uses: actions/checkout@v1
+    - name: Install ganache-cli
+      run: sudo npm install -g ganache-cli
+    - name: Run the beacon chain sim without an eth1 connection
+      run: cargo run --release --bin simulator no-eth1-sim
